refactor: split generated run command repositories
